fix: add new module constants. Remove unused code
