+++ js/create-location.js
@@ -1,5 +1,5 @@
 import {getRandomFloatNumber} from './util.js';
-import {MIN_LATITUDE, MIN_LONGITUDE, MAX_LATITUDE, MAX_LONGITUDE} from './create-offer.js';
+import {MIN_LATITUDE, MIN_LONGITUDE, MAX_LATITUDE, MAX_LONGITUDE} from './constants.js';
 
 const createLocation = () => ({
   lat: getRandomFloatNumber(MIN_LATITUDE, MAX_LATITUDE),
