+++ js/create-offer.js
@@ -1,4 +1,5 @@
 import {getRandomInteger, getRandomFloatNumber, getRandomArrayElement, getRandomShuffleArray} from './util.js';
+import {TYPES, CHECK_TIME, FEATURES, MIN_LATITUDE, MAX_LATITUDE, MIN_LONGITUDE, MAX_LONGITUDE} from './constants.js';
 
 const TITLES = [
   'Апартаменты просторные для большой семьи',
@@ -11,29 +12,6 @@ const TITLES = [
   'Светлая студия рядом с Сенной площадью',
   'Уютные апартаменты в историческом центре города',
   'Двухкомнатная квартира с террасой в стиле Сканди'
-];
-
-const TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel'
-];
-
-const CHECK_TIME = [
-  '12:00',
-  '13:00',
-  '14:00'
-];
-
-const FEATURES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner'
 ];
 
 const DESCRIPTIONS = [
@@ -59,10 +37,6 @@ const MIN_SYMBOL = 1;
 const MAX_SYMBOL = 10;
 const MIN_PRICE = 10000;
 const MAX_PRICE = 50000;
-const MIN_LATITUDE = 35.65000;
-const MAX_LATITUDE = 35.70000;
-const MIN_LONGITUDE = 139.70000;
-const MAX_LONGITUDE = 139.80000;
 
 const createOffer = () => ({
   title: getRandomArrayElement(TITLES),
