+++ js/main.js
@@ -1,7 +1,3 @@
-import './create-author.js';
-import './create-location.js';
-import './create-offer.js';
-import './util.js';
 import {similiarAnnounce} from './data.js';
 
 similiarAnnounce();
